fix compiler warning: change lower_name param to size_t
git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@1574 c4029f9c-b713-0410-a448-fae9e31a314a

+++ dnssec-tools/lib/libval/val_support.h
@@ -96,7 +96,7 @@ int extract_from_rr (   u_int8_t *response,
                         u_int32_t *ttl_h,
                         u_int16_t *rdata_length_h,
                         int *rdata_index);
-void lower_name (u_int8_t rdata[], int *index);
+void lower_name (u_int8_t rdata[], size_t *index);
 void lower (u_int16_t type_h, u_int8_t *rdata, int len);
 struct rr_rec *copy_rr_rec (u_int16_t type_h, struct rr_rec *r, int dolower);
 int link_rr (struct rr_rec **cs, struct rr_rec *cr);
